Update for Zeek plugin reorg.
This is copying over a bunch of cmake/testing infrastructure from the
new spicy-plugin repository. Also adapting to the plugin's new
standalone build, and adding support for installation outside of zkg.

We may eventually factor some of the cmake/testing logic out and have it
shared across the two packages (and potentially other ones as well).