Update for plugin reorganisation.
Also aligning scripts that are shared with the plugin so that both are
using identical versions. This is in preparation for the analyzers
package to later switch to using the versions provided by the plugin
directly.

Also fixing a version issue where scripts for the ZIP analyzer could end
up being loaded even when we didn't compile the analyzer.