stun: Fix integer overflow/underflow issues
The attribute length integer overflow was seen in the wild. The truncated
error attribute length was constructed artificially after light inspection
of the grammar.

Test pcaps are based on the stun-ice-testcall.pcap and modified with
scapy to force the attribute length to 0xffff or truncate the length
down to 0x0003 for the error attribute.

Main motivation for these fixes: The integer overflow is logged as
AnalyzerError() and makes the Zeek logs noisy due to these errors
being printed to stderr:

    analyzer/stun/stun.spicy, line 165: integer overflow